v3

+++ sort.py
@@ -12,11 +12,11 @@ video = ".avi .mp4 .mov .mkv".split()
 archives = ".zip .gz .tar".split()
 
 dict_suffixes = {
-    'documents': [ext.lower() for ext in documents],
-    'images': [ext.lower() for ext in images],
-    'audio': [ext.lower() for ext in audio],
-    'video': [ext.lower() for ext in video],
-    'archives': [ext.lower() for ext in archives],
+    'documents' : documents,
+    'images' : images,
+    'audio' : audio,
+    'video' : video,
+    'archives' : archives,
 }
 
 dict_suffixes_reverse = {}
@@ -69,29 +69,18 @@ def move_file(root_path: Path, path_file: Path):
     name = normalize(path_file.stem)
     suff = path_file.suffix.lower()
     
-    if path_file.is_file():        
-        category = dict_suffixes_reverse.get(suff, 'others')
-        new_path_dir = root_path / category
+    category = dict_suffixes_reverse.get(suff, 'others')
 
-        if not new_path_dir.exists():
-            new_path_dir.mkdir()
+    new_path_dir = root_path/category
 
-        new_path_file = new_path_dir / (name + suff)
-        path_file.replace(new_path_file)
+    if not new_path_dir.exists():
+        new_path_dir.mkdir()
 
-        if category == 'archives':
-            unpack(new_path_file, new_path_dir)
+    new_path_file = new_path_dir / (name + suff)
+    path_file.replace(new_path_file)
 
-    elif path_file.is_dir():        
-        new_path_dir = root_path / 'others'
-        if not new_path_dir.exists():
-            new_path_dir.mkdir()
-
-        new_path_dir = new_path_dir / name
-        path_file.replace(new_path_dir)
-
-    else:
-        return
+    if category == 'archives':
+        unpack(new_path_file, new_path_dir)
 
 def sort_folder(root_path: Path, path: Path):  
     all_files = []
@@ -110,8 +99,9 @@ def sort_folder(root_path: Path, path: Path):
             move_file(root_path, item)
 
         if item.is_dir():
-            sort_folder(root_path, item)
-            item.rmdir()
+            if not(item.parts[-1] in dict_suffixes and item.parts[:-1] == root_path.parts):
+                sort_folder(root_path, item)
+                item.rmdir()
             
 
     write_list_to_file(root_path / 'all_files.txt', all_files)
